setting loading blur component

+++ server/routes/upload_profile_image.js
@@ -15,8 +15,8 @@ const storage = multer.diskStorage({
 		cb(null, 'images')
 	},
 	filename: (req, file, cb) => {
-		newImageName = Date.now() + "_" + 
-			crypto.createHash('md5').update(file.originalname).digest('hex') + 
+		newImageName = Date.now() + "_" +
+			crypto.createHash('md5').update(file.originalname).digest('hex') +
 			path.extname(file.originalname)
 		cb(null, newImageName)
 	},
@@ -29,7 +29,7 @@ const upload = util.promisify(
 		fileFilter: (req, file, cb) => {
 			if (path.extname(file.originalname) != '.jpg' && path.extname(file.originalname) != '.png' && path.extname(file.originalname) != '.jpeg') {
 				return cb("Invalid file type, try uploading a '.jpg', '.jpeg' or a '.png' file")
-			} else { 
+			} else {
 				cb(null, true)
 			}
 		}
@@ -38,11 +38,12 @@ const upload = util.promisify(
 
 router.post('/', validateToken, async (req, res) => {
 	try {
+		if (!fs.existsSync('./images')) // create the folder /images if it doesn't exist
+			fs.mkdirSync('./images')
 		var result = await queryPromise( // check if a profile picture already exists, to remplace it
 			"SELECT * FROM images WHERE uid = ? AND isProfileImage = 1",
 			[req.user.id]
 		)
-		console.log('length=' + result.length)
 		await upload(req, res)
 		if (result.length == 0) { // if no profile image was added add one
 			await queryPromise(
@@ -50,11 +51,13 @@ router.post('/', validateToken, async (req, res) => {
 				[req.user.id, 1, newImageName],
 			)
 		} else { // else if a profile image exists
-			fs.unlinkSync(`./images/${result[0].image}`) // delete the image from ./images/
-			await queryPromise(
-				"UPDATE images SET image = ? WHERE id = ?",
-				[newImageName, result[0].id]
-			)
+			if (fs.existsSync(`./images/${result[0].image}`)) { // we know the img exists in the DB, but check if the file exists in the /images folder or has been deleted manually
+				fs.unlinkSync(`./images/${result[0].image}`) // delete the image from ./images/
+				await queryPromise(
+					"UPDATE images SET image = ? WHERE id = ?",
+					[newImageName, result[0].id]
+				)
+			}
 		}
 		res.send("Profile image upladed successfully")
 	} catch (err) {
